pyproject.toml
[project]
name = "fastapi-replicate-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aerich[toml]>=0.9.0",
    "argon2-cffi>=23.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "ruff>=0.11.12",
    "setuptools>=80.9.0",
    "tortoise-orm[asyncpg]>=0.25.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "fastapi-replicate-admin",
]

[tool.uv.sources]
fastapi-replicate-admin = { workspace = true }
